Fix scrollbar being incorrectly positioned when first shown
Previously, if the scrollbar were initially hidden (due to having a small amount of content) and then enough content was added that the scrollbar became visible, the scrollbar would be a few pixels too far to the right. This fixes that issue by computing the 'right' property after the 'ss-hidden' class is removed.